flask_backend/backend/requirements.txt
absl-py==1.0.0
astunparse==1.6.3
beautifulsoup4==4.10.0
blinker==1.4
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
cryptography==36.0.1
Flask==2.0.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
flatbuffers==2.0
gast==0.4.0
google==3.0.0
google-api-core==2.4.0
google-api-python-client==2.36.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.54.0
greenlet==1.1.2
grpcio==1.43.0
h5py==3.6.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
# jwt==1.3.1
keras==2.7.0
Keras-Preprocessing==1.1.2
libclang==12.0.0
Markdown==3.3.6
MarkupSafe==2.0.1
# nltk==3.6.7
# numpy==1.22.1
oauthlib==3.1.1
opt-einsum==3.3.0
protobuf==3.19.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
# PyJWT==2.3.0
pyparsing==3.0.6
regex==2022.1.18
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.30
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
# tensorflow==2.7.0
# tensorflow-estimator==2.7.0
# tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
tqdm==4.62.3
typing-extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.8
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.7.0
